requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
aiohttp>=3.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
rich>=13.7.0
apscheduler>=3.10.0
requests>=2.31.0
